+++ websocket-debugger/src/main/kotlin/at/jku/ssw/wsdebug/communication/DebugWebSocketServer.kt
@@ -1,6 +1,9 @@
 package at.jku.ssw.wsdebug.communication
 
+import at.jku.ssw.wsdebug.DEVELOPMENT_MODE
 import at.jku.ssw.wsdebug.asStringWithStackTrace
+import at.jku.ssw.wsdebug.sendAndPrintResponse
+import at.jku.ssw.wsdebug.shorten
 import org.java_websocket.WebSocket
 import org.java_websocket.handshake.ClientHandshake
 import org.java_websocket.server.WebSocketServer
@@ -14,8 +17,6 @@ class DebugWebSocketServer(address: InetSocketAddress) : WebSocketServer(address
     override fun onOpen(conn: WebSocket?, handshake: ClientHandshake?) {
         println("[WebSocket] Event: onOpen")
         println("  Connection with client is now established, ready to process requests")
-        println("  Connection: $conn")
-        println("  Handshake: $handshake")
         println()
         // only one debugging session / connection at a time, new ones preempt old ones
         activeConnection?.close()
@@ -25,7 +26,6 @@ class DebugWebSocketServer(address: InetSocketAddress) : WebSocketServer(address
 
     override fun onClose(conn: WebSocket, code: Int, reason: String, remote: Boolean) {
         println("[WebSocket] Event: onClose")
-        println("  Connection: $conn")
         println("  Code: $code")
         println("  Reason: $reason")
         println("  Remote: $remote")
@@ -35,8 +35,7 @@ class DebugWebSocketServer(address: InetSocketAddress) : WebSocketServer(address
 
     override fun onMessage(conn: WebSocket, message: String) {
         println("[WebSocket] Event: onMessage")
-        println("  Connection: $conn")
-        println("  Message: $message")
+        println("  Message: ${if (DEVELOPMENT_MODE) message else message.shorten(100)}")
 
         thread(isDaemon = true, name = "Handler for $message") {
             conn.sendAndPrintResponse(generateResponse(message), 100)
@@ -52,7 +51,6 @@ class DebugWebSocketServer(address: InetSocketAddress) : WebSocketServer(address
 
     override fun onError(conn: WebSocket?, ex: Exception) {
         println("[WebSocket] Event: onError")
-        println("  Connection: $conn")
         println("  Exception: ${ex.asStringWithStackTrace()}")
         println()
     }